episode-eight/requirements_updated.txt
aiofiles==22.1.0
aiosqlite==0.18.0
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==22.2.0
aws-cdk-lib==2.88.0
aws-cdk.asset-awscli-v1==2.2.200
aws-cdk.asset-kubectl-v20==2.1.2
aws-cdk.asset-node-proxy-agent-v5==2.0.166
aws-lambda-powertools==1.30.0
aws-solutions-constructs.aws-lambda-dynamodb==2.41.0
aws-solutions-constructs.core==2.41.0
aws-xray-sdk==2.11.0
Babel==2.12.1
backcall==0.2.0
bandit==1.7.5
beautifulsoup4==4.12.0
black==23.1.0
bleach==6.0.0
blessed==1.20.0
boto3==1.26.98
botocore==1.29.98
CacheControl==0.12.11
cachy==0.3.0
cattrs==22.2.0
cdk-nag==2.23.5
certifi==2022.12.7
cffi==1.15.1
chalice==1.28.0
charset-normalizer==3.1.0
cleo==0.8.1
click==8.1.3
clikit==0.6.2
comm==0.1.3
constructs==10.1.289
coverage==7.2.2
crashtest==0.3.1
cryptography==40.0.1
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
exceptiongroup==1.1.1
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.8.0
flake8==6.0.0
fpdf==1.7.2
fqdn==1.5.1
freezegun==1.2.2
gitdb==4.0.10
GitPython==3.1.31
html5lib==1.1
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
iniconfig==2.0.0
inquirer==2.10.1
ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
jsii==1.85.0
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter-client==8.1.0
jupyter-core==5.3.0
jupyter-events==0.6.3
jupyter-server==2.5.0
jupyter-server-fileid==0.8.0
jupyter-server-terminals==0.4.4
jupyter-server-ydoc==0.8.0
jupyter-ydoc==0.2.3
jupyterlab==3.6.2
jupyterlab-pygments==0.2.2
jupyterlab-server==2.21.0
keyring==23.8.2
lockfile==0.12.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mistune==2.0.5
moto==4.1.6
msgpack==1.0.4
mypy-extensions==1.0.0
nbclassic==0.5.4
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.8.0
nest-asyncio==1.5.6
notebook==6.5.3
notebook-shim==0.2.2
numpy==1.23.3
packaging==23.0
pandas==1.5.0
pandocfilters==1.5.0
parso==0.8.3
pastel==0.2.1
pathspec==0.11.1
patsy==0.5.3
pbr==5.11.1
pexpect==4.8.0
pickleshare==0.7.5
pkginfo==1.8.3
pkgutil-resolve-name==1.3.10
platformdirs==3.1.1
pluggy==1.0.0
poetry==1.1.15
poetry-core==1.0.8
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
publication==0.0.3
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
Pygments==2.14.0
pylev==1.4.0
pyparsing==3.0.9
pyrsistent==0.19.3
pytest==7.2.2
pytest-cov==4.0.0
pytest-mock==3.10.0
python-dateutil==2.8.2
python-editor==1.0.4
python-json-logger==2.0.7
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
readchar==4.0.5
requests==2.28.2
requests-toolbelt==0.9.1
responses==0.23.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.2
s3transfer==0.6.0
scipy==1.10.1
Send2Trash==1.8.0
shellingham==1.5.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4
stack-data==0.6.2
statsmodels==0.13.2
stevedore==5.0.0
terminado==0.17.1
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.11.4
tornado==6.2
traitlets==5.9.0
typeguard==2.13.3
types-PyYAML==6.0.12.9
typing-extensions==4.5.0
tzdata==2023.3
uri-template==1.2.0
urllib3==1.26.15
virtualenv==20.16.3
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
wrapt==1.15.0
xmltodict==0.13.0
y-py==0.5.9
ypy-websocket==0.8.2
zipp==3.15.0
